move action to catalog service where it belongs
