+++ lib/shared/addon/catalog/service.js
@@ -3,6 +3,9 @@ import Service, { inject as service } from '@ember/service';
 import { addQueryParams, uniqKeys } from 'shared/utils/util';
 import C from 'shared/utils/constants';
 import EmberObject from '@ember/object'
+import { get } from '@ember/object';
+import { /* parseExternalId, */ parseHelmExternalId } from 'ui/utils/parse-externalid';
+import { all } from 'rsvp';
 
 const RANCHER_VERSION = 'rancherVersion';
 const DEFAULT_BASE = 'kubernetes';
@@ -34,6 +37,17 @@ export default Service.extend({
         url: `${this.get('app.apiEndpoint')}/catalogs?action=refresh`,
       });
     });
+  },
+
+  fetchAppTemplates: function(apps) {
+    let deps      = [];
+
+    apps.forEach((app) => {
+      let extInfo = parseHelmExternalId(app.get('externalId'));
+      deps.push(this.fetchTemplate(extInfo.templateId, false));
+    });
+
+    return all(deps);
   },
 
   fetchCatalogs(opts) {
